Upgrade performance of SeqAlloc, it is now ~40 times faster.
This uses the numpy.ctypes interface to create a void pointer directly. Therefore we avoid any python loops. On my system, projecting a single pattern went down from 178 ms to 4 ms.

I have visually checked that this gives the same result, even if the order of the numpy array is 'F' instead of 'C'. If you know a better way to check please let me know.